Move LibraryProps to Library file
