+++ src/content/LibraryContent.tsx
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { LibraryProps } from '../theme/types';
+import { LibraryProps } from '../theme/components';
 
 export const libraries: LibraryProps[] = [
   {
